наладил свзь фронт бек, сделал первый роут связанный с проектами и тасками

+++ app/src/main/java/com/nurlan1507/task_manager_mobile/ui_components/main_screen/bottom_sheet_layouts/TaskCreationBottomSheetLayout.kt
@@ -157,7 +157,7 @@ fun TaskCreationBottomSheetLayout(tasksViewModel: TasksViewModel, sheetState: Mo
             Row(modifier = Modifier
                 .fillMaxWidth()
                 .height(40.dp), horizontalArrangement = Arrangement.SpaceBetween){
-                ProjectSelectionButton(projectId = tasksViewModel.fieldState.value.projectId, projectList = tasksState.value.projects){projectId->
+                ProjectSelectionButton(projectId = tasksViewModel.fieldState.value.projectId.toInt(), projectList = tasksState.value.projects){projectId->
                     tasksViewModel.onEvent(TasksEvent.EnteredProjectId(projectId))
                 }
                 Box(modifier = Modifier
